Shuffle namespaces and make hitcircle pieces reusable by other objects.
