+++ osu.Game.Mode.Osu/Objects/Drawables/DrawableSlider.cs
@@ -1,0 +1,17 @@
+﻿using osu.Game.Modes.Objects.Drawables;
+
+namespace osu.Game.Modes.Osu.Objects.Drawables
+{
+    class DrawableSlider : DrawableHitObject
+    {
+        public DrawableSlider(Slider h) : base(h)
+        {
+
+        }
+
+        protected override void UpdateState(ArmedState state)
+        {
+
+        }
+    }
+}
