+++ osu.Game.Mode.Osu/Objects/Drawables/Pieces/GlowPiece.cs
@@ -1,0 +1,36 @@
+﻿using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Textures;
+
+namespace osu.Game.Modes.Osu.Objects.Drawables.Pieces
+{
+    public class GlowPiece : Container
+    {
+        private Sprite layer;
+
+        public GlowPiece()
+        {
+            Anchor = Anchor.Centre;
+            Origin = Anchor.Centre;
+
+            Children = new[]
+            {
+                layer = new Sprite
+                {
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    BlendingMode = BlendingMode.Additive,
+                    Alpha = 0.5f
+                }
+            };
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(TextureStore textures)
+        {
+            layer.Texture = textures.Get(@"Play/osu/ring-glow@2x");
+        }
+    }
+}
